Move out send cloud request
